Added a database handler class for using MySQL connection pools and handling common errors rather than the system shutting down.
Added error checking to all database requests to detect valid connections and cursors

Changed the http request library from httplib to requests and added exception handling for when the relay controller
doesn't respond as expected.

Added detection for failed remote sensors so the sensor is removed from the selector list if it has failed and is
displayed again when it is working again.